Fixes #2721 - Move custom views to separate files
