+++ Blockzilla/PhotonActionSheetCell.swift
@@ -5,12 +5,6 @@
 import UIKit
 import SnapKit
 
-extension UIScreen {
-    var isSmallScreen: Bool {
-        return min(bounds.size.width, bounds.size.height) < 700
-    }
-}
-
 struct PhotonActionSheetCellUX {
     static let LabelColor = UIColor.blue
     static let BorderWidth: CGFloat = CGFloat(0.5)
