+++ Blockzilla/SettingsViewController.swift
@@ -10,82 +10,6 @@ import Intents
 import IntentsUI
 import Glean
 
-class SettingsTableViewCell: UITableViewCell {
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        selectionStyle = .gray
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    internal func setupDynamicFont(forLabels labels: [UILabel], addObserver: Bool = false) {
-        for label in labels {
-            label.font = UIFont.preferredFont(forTextStyle: UIFont.TextStyle.body)
-            label.adjustsFontForContentSizeCategory = true
-        }
-
-        NotificationCenter.default.addObserver(forName: UIContentSizeCategory.didChangeNotification, object: nil, queue: nil) { _ in
-            self.setupDynamicFont(forLabels: labels)
-        }
-    }
-}
-
-class SettingsTableViewAccessoryCell: SettingsTableViewCell {
-    var labelText: String? {
-        get { return textLabel?.text }
-        set {
-            textLabel?.text = newValue
-        }
-    }
-    
-    var accessoryLabelText: String? {
-        get { return detailTextLabel?.text }
-        set {
-            detailTextLabel?.text = newValue
-        }
-    }
-
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        accessoryView = UIImageView(image: UIImage(systemName: "chevron.right"))
-        tintColor = .secondaryText.withAlphaComponent(0.3)
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-class SettingsTableViewToggleCell: SettingsTableViewCell {
-    private let newLabel = SmartLabel()
-    private let spacerView = UIView()
-    var navigationController: UINavigationController?
-
-    init(style: UITableViewCell.CellStyle, reuseIdentifier: String?, toggle: BlockerToggle) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        setupDynamicFont(forLabels: [newLabel], addObserver: true)
-
-        newLabel.numberOfLines = 0
-        newLabel.text = toggle.label
-
-        textLabel?.numberOfLines = 0
-        textLabel?.text = toggle.label
-
-        newLabel.textColor = .primaryText
-        textLabel?.textColor = .primaryText
-        layoutMargins = UIEdgeInsets.zero
-
-        accessoryView = PaddedSwitch(switchView: toggle.toggle)
-        selectionStyle = .none
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
 class SettingsViewController: UIViewController, UITableViewDataSource, UITableViewDelegate {
     enum Section: String {
         case general, privacy, usageData, studies, search, siri, integration, mozilla
