+++ Blockzilla/SmartLabel.swift
@@ -15,10 +15,3 @@ class SmartLabel: UILabel {
         setupShrinkage()
     }
 }
-
-extension UILabel {
-    func setupShrinkage() {
-        adjustsFontSizeToFitWidth = true
-        minimumScaleFactor = 0.6 // Arbitrarily set to 60%
-    }
-}
