+++ Blockzilla/UIPasteBoardExtensions.swift
@@ -1,0 +1,30 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+import UIKit
+
+extension UIPasteboard {
+
+    //
+    // As of iOS 11: macOS/iOS's Universal Clipboard feature causes UIPasteboard to block.
+    //
+    // (Known) Apple Radars that have been filed:
+    //
+    //  http://www.openradar.me/28787338
+    //  http://www.openradar.me/28774309
+    //
+    // Discussion on Twitter:
+    //
+    //  https://twitter.com/steipete/status/787985965432369152
+    //
+    //  To workaround this, urlAsync(callback:) makes a call of UIPasteboard.general on
+    //  an async dispatch queue, calling the completion block when done.
+    //
+    func urlAsync(callback: @escaping (URL?) -> Void) {
+        DispatchQueue.global().async {
+            let url = URL(string: UIPasteboard.general.string ?? "")
+            callback(url)
+        }
+    }
+}
