+++ Blockzilla/URLExtensions.swift
@@ -57,6 +57,21 @@ private var etldEntries: TLDEntryMap? = {
 
 // MARK: - Local Resource URL Extensions
 extension URL {
+    
+    /// Construct an URL pointing to a specific topic on SUMO. The topic comes from the Topics enum.
+    ///
+    /// The resulting URL will include the app version, operating system and locale code. For example, a topic
+    /// "cheese" will be turned into a link that looks like https://support.mozilla.org/1/mobile/2.0/iOS/en-US/cheese
+    ///
+    /// If for some reason the URL could not be created, a default URL to support.mozilla.org is returned. This is
+    /// a very rare case that should not happen except in the rare case where the URL may be dynamically formatted.
+    init(forSupportTopic topic: SupportTopic) {
+        guard let escapedTopic = topic.slug.addingPercentEncoding(withAllowedCharacters: CharacterSet.urlPathAllowed), let languageIdentifier = Locale.preferredLanguages.first else {
+            self.init(string: SupportTopic.fallbackURL)!
+            return
+        }
+        self.init(string: "https://support.mozilla.org/1/mobile/\(AppInfo.shortVersion)/iOS/\(languageIdentifier)/\(escapedTopic)")!
+    }
 
     public func allocatedFileSize() -> Int64 {
         // First try to get the total allocated size and in failing that, get the file allocated size
@@ -253,6 +268,14 @@ extension URL {
     public func isWebPage(includeDataURIs: Bool = true) -> Bool {
         let schemes = includeDataURIs ? ["http", "https", "data"] : ["http", "https"]
         return scheme.map { schemes.contains($0) } ?? false
+    }
+    
+    public func isWebPage() -> Bool {
+        let schemes = ["http", "https"]
+        if let scheme = scheme, schemes.contains(scheme) {
+            return true
+        }
+        return false
     }
 
     // This helps find local urls that we do not want to show loading bars on.
